+++ src/main/java/team/cagayakegirls/mafglib/utils/ModPlatform.java
@@ -1,5 +1,6 @@
 package team.cagayakegirls.mafglib.utils;
 
+import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.loading.FMLLoader;
 import net.neoforged.fml.loading.FMLPaths;
@@ -7,6 +8,7 @@ import net.neoforged.neoforgespi.language.IModInfo;
 
 import java.nio.file.Path;
 import java.util.List;
+import java.util.Optional;
 
 public class ModPlatform {
     public static boolean isModLoaded(String modId) {
@@ -23,5 +25,9 @@ public class ModPlatform {
 
     public static List<IModInfo> getAllMods() {
         return ModList.get().getMods();
+    }
+
+    public static Optional<? extends ModContainer> getModContainer(String modId) {
+        return ModList.get().getModContainerById(modId);
     }
 }