+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## Changelog
-- sync upstream [`1.21.10-0.26.4`](https://github.com/sakura-ryoko/malilib/releases/tag/1.21.10-0.26.4) change
-- update NeoForge version `21.10.33-beta` -> `21.10.52-beta`
+- add `ModPlatform.getModContainer`